Fix: correct company index import path
