+++ src/validators/company/types.ts
@@ -1,0 +1,7 @@
+import { AttributeType } from "../attribute-validator.js";
+export interface CachedTypeInfo {
+  fieldType: string;
+  attioType: string;
+  validatorType: AttributeType;
+  timestamp: number;
+}
